+++ hypervisor/include/jailhouse/printk.h
@@ -13,7 +13,7 @@
 #include <jailhouse/types.h>
 
 extern volatile unsigned long panic_in_progress;
-extern unsigned int panic_cpu;
+extern unsigned long panic_cpu;
 
 void printk(const char *fmt, ...);
